Move getProjectModels to AbstractGraphResource.
